# perf: do not process nodes multiple times