Also, use `vertex.dist` instead of looking up in map.

The effect is smaller than I anticipated, but still around 5%.

```
$ benchstat /tmp/ch-bench/master-10-1.txt /tmp/ch-bench/processed-check-10-1.txt
goos: linux
goarch: amd64
pkg: github.com/LdDl/ch
cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
                                                             │ /tmp/ch-bench/master-10-1.txt │ /tmp/ch-bench/processed-check-10-1.txt │
                                                             │            sec/op             │     sec/op       vs base               │
ShortestPath/CH_shortest_path/4/vertices-4-edges-9-6                             1.356µ ± 3%       1.334µ ± 2%       ~ (p=0.305 n=10)
ShortestPath/CH_shortest_path/8/vertices-8-edges-61-6                            2.576µ ± 9%       2.483µ ± 1%  -3.65% (p=0.000 n=10)
ShortestPath/CH_shortest_path/16/vertices-16-edges-316-6                         8.795µ ± 3%       8.208µ ± 4%  -6.67% (p=0.000 n=10)
ShortestPath/CH_shortest_path/32/vertices-32-edges-1404-6                        21.24µ ± 6%       19.91µ ± 1%  -6.27% (p=0.000 n=10)
ShortestPath/CH_shortest_path/64/vertices-64-edges-5894-6                        51.32µ ± 2%       49.15µ ± 2%  -4.24% (p=0.002 n=10)
ShortestPath/CH_shortest_path/128/vertices-128-edges-23977-6                     140.0µ ± 2%       135.0µ ± 3%  -3.55% (p=0.000 n=10)
ShortestPath/CH_shortest_path/256/vertices-256-edges-97227-6                     354.9µ ± 4%       333.1µ ± 2%  -6.15% (p=0.000 n=10)
geomean                                                                          20.76µ            19.81µ       -4.61%
```